feat: allow multiple read/write on single disk image
This feature is extremely useful for running multiple QFlex instances
sharing the same qcow2 disk file in parallel. It does not acquire the
write lock of the qcow2 file anymore. Instead, it creates an overlay
using the specified disk as its backing file (snapshot=on), adjust the
backing file to the point of a given snapshot (tmp-snapshot-name=XXX).